HW 6.3 end

+++ src/lesson6/HW6/Animals.java
@@ -24,7 +24,7 @@ public abstract class Animals {
         this.maxRun = maxRun;
     }
 
-    public void setMaxJump(double maxJump) {
+    public void setMaxJump(int maxJump) {
         this.maxJump = maxJump;
     }
